# [pyper] to + lengths_to_offsets